+++ plugin/src/helpers/constants.ts
@@ -1,0 +1,47 @@
+// Constants for color values
+export const COLORS = {
+  base: "#F9F7FD",
+  text: "#111111",
+  link: "#0000FF",
+  stroke: "#E3D5F6",
+  dark: "#111111",
+  highlight: "#7938D3",
+  white: "#FFFFFF"
+};
+
+
+
+// Constants for font weights
+export const FONT_WEIGHT = {
+  regular: "Regular",
+  bold: "Bold"
+};
+
+// Constants for font sizes
+export const FONT_SIZES = {
+  large: 28,
+  medium: 20,
+  small: 16
+};
+
+// Constants for line heights
+export const LINE_HEIGHTS = {
+  large: 36,
+  medium: 28,
+  small: 24
+};
+
+// Padding sizes
+export const PADDING = {
+  large: 16,
+  medium: 8,
+  small: 4,
+  none: 0
+};
+
+// Corner radius
+export const CORNER_RADIUS = {
+  large: 12,
+  medium: 8,
+  small: 4
+};
