+++ plugin/webpack.config.js
@@ -1,0 +1,37 @@
+const path = require('path');
+const HtmlWebpackPlugin = require('html-webpack-plugin');
+
+module.exports = {
+  entry: './src/index.ts',
+  output: {
+    path: path.resolve(__dirname, 'build'),
+    filename: 'code.js'
+  },
+  resolve: {
+    extensions: ['.ts', '.js']
+  },
+  module: {
+    rules: [
+      {
+        test: /\.ts$/,
+        use: {
+          loader: 'ts-loader',
+          options: {
+            happyPackMode: true,
+            configFile: 'tsconfig.json'  
+          }
+        },
+        exclude: /node_modules/
+      }
+    ]
+  },
+  plugins: [
+    new HtmlWebpackPlugin({
+      template: './src/index.html',
+      filename: 'ui.html',
+      inject: false
+    })
+  ],
+  mode: 'production', // Consider using 'production' for final builds
+  devtool: false
+};
